refactor: move to ext folder and add credit
